refactor(fc-invoke): split substrate into cluster/ + node/ planes (ADR 031)
Draw the control-plane / data-plane boundary in the agent substrate now, while it
is still an in-process Go interface, so the future central-agent + per-node
DaemonSet split is a wiring change rather than a rewrite. Pure move-and-rewire:
one binary, one Deployment, no behavior change.

- cluster/ (control plane): ingress (was invoke/internal/server, the HTTP front
  that routes /invoke/{workload} to a node executor) + catalog (was
  invoke/internal/config, the workload table loader). The workload-to-executor
  map is the trivial single-node placement table; a placement + fleet package
  (scaling, in-place pod resize) slot in here later.
- node/ (data plane): invoker, fcvm (driver + fcclient), vsockhttp, egress, all
  moved under node/ unchanged.
- substrate/ holds the neutral seam both planes depend on: the new NodeExecutor
  interface (was server.Invoker) and the Workload spec type (moved out of the
  catalog so the node plane consumes it without importing the cluster plane). The
  dependency graph is now cluster -> substrate <- node; neither plane imports the
  other. invoke/cmd is the only package that wires both.

The image binary (invoke/cmd:invoke) and the chart are untouched. ADR 031 records
the direction and the cmd/cluster-agent + cmd/node-daemon + remoteNode future.
