+++ docs/decisions/index.md
@@ -36,6 +36,7 @@ ADRs document significant architectural decisions and their context.
 | [027 - Agent GitHub App Roles: Implementer and Reviewer](agents/027-agent-github-app-roles.md)                                                  | Split the agent's single `gh` identity into two GitHub Apps mapped to roles: a `claude/*`-scoped, no-merge implementer (lower-trust model) and an approve-and-merge reviewer (Opus or better, adversarial spec-vs-impl). Merge gated by a reviewer-controlled `agent-review/gate` status check since App bots cannot be CODEOWNERS; `@jomcgi` stays the human owner. Role token injected per-thread via the 023 egress swap. Draft                                                                                                                                                                                                                                                                                                                                                                    |
 | [028 - Elastic Agent-MicroVM Capacity and State-Preserving Reclaim](agents/028-elastic-agent-microvm-capacity-and-reclaim.md)                   | Run the disposable agent-microVM tier elastically on a shared node: reaffirm FC-direct (sub-100ms golden-template restore; kata-fc has no CRI snapshot); honest burst via dynamically-sized Guaranteed in-place pod resize + a headroom watermark controller (k8s off the hot path, `Infeasible` resize = spill signal); critical-tenant-safe reclaim via an idle-first victim ladder + graceful drain (checkpoint -> park -> rehydrate) driven by a free-allocatable cushion + node-pinned-Pending watch; and a filesystem-as-durability-contract state model (`/repos` git branches to a state remote, `/data` to S3, rest recomputed) checkpointed as a consistent tuple. Drops per-thread VM snapshots; ballast pods + sharding considered and deferred. Draft                                    |
 | [030 - fc-invoke, a Single Configurable Surface for Running Workloads in Firecracker](agents/030-fc-invoke-configurable-firecracker-surface.md) | One host daemon that implements ADR 019's stubbed `Substrate.Exec` seam plus the orchestration around it, absorbing `semgrep-scand` and `fc-agentd`'s VM lifecycle. HTTP ingress -> HTTP-over-vsock reverse proxy into a guest; workloads are named Helm-values entries (seven generic knobs: image/resources/concurrency/egress/warmBase/sessioned/requestTimeout); warm-base (platform VM snapshot) and state hydration (guest-side shim capabilities) split on the litmus line; all durable state owned by orchestrators (stateless daemon, 2-way door); a shared bazel-baked guest shim (HTTP server + git/object-store capabilities) keeps the rich guest toolkit out of the thin host contract. Home `projects/firecracker/{substrate,goosecracker,semgrep}`. Supersedes ADR 025 in part. Draft |
+| [031 - Control-Plane / Data-Plane Split for the Agent Substrate (cluster + node)](agents/031-cluster-node-control-data-plane-split.md)          | Split the substrate Go code into a `cluster/` control plane (ingress, catalog, future placement + fleet/in-place-resize) and a `node/` data plane (invoker, fcvm driver, vsockhttp, egress), bracketed by a neutral `substrate.NodeExecutor` seam so neither plane imports the other. One binary/Deployment now (in-process local executor); the physical split to a central agent plus per-node DaemonSet later is a wiring change behind the seam, not a rewrite. Accepted                                                                                                                                                                                                                                                                                                                          |
 
 ## Docs
