+++ projects/firecracker/substrate/cluster/catalog/config_test.go
@@ -1,4 +1,4 @@
-package config
+package catalog
 
 import (
 	"os"
