--- projects/firecracker/substrate/invoke/internal/config/BUILD
@@ -1,14 +1,0 @@
-load("@rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "config",
-    srcs = ["config.go"],
-    importpath = "github.com/jomcgi/homelab/projects/firecracker/substrate/invoke/internal/config",
-    visibility = ["//projects/firecracker/substrate/invoke:__subpackages__"],
-)
-
-go_test(
-    name = "config_test",
-    srcs = ["config_test.go"],
-    embed = [":config"],
-)
