+++ projects/firecracker/substrate/substrate/BUILD
@@ -3,8 +3,10 @@ load("@rules_go//go:def.bzl", "go_library", "go_test")
 go_library(
     name = "substrate",
     srcs = [
+        "executor.go",
         "fake.go",
         "substrate.go",
+        "workload.go",
     ],
     importpath = "github.com/jomcgi/homelab/projects/firecracker/substrate/substrate",
     visibility = ["//visibility:public"],
